requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
boto3==1.24.49
botocore==1.27.49
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
urllib3==1.26.11
zope.interface==5.4.0
